requirements.txt
# torch==2.1.2
# torchvision==0.16.2
# torchaudio==2.1.2
accelerate==0.25.0
appdirs==1.4.4
loralib==0.1.2
bitsandbytes==0.41.3.post2
datasets==2.15.0
fire==0.5.0
peft==0.7.1
transformers==4.38.2
# transformers==4.35.2
# transformers==4.36.2
sentencepiece==0.1.99
py7zr==0.20.8
scipy==1.11.4
optimum==1.16.1
mosaicml-streaming==0.7.2
wandb>=0.16.1